+++ History.md
@@ -1,6 +1,7 @@
 1.0.16 / 2015-??-??
 ==================
-  * Add file-instrumentation only target `jscover-file:instrument` (https://github.com/tntim96/JSCover-maven-plugin/issues/14)
+  * Add file-instrumentation only goal `jscover-file:instrument` (https://github.com/tntim96/JSCover-maven-plugin/issues/14)
+  * Add merge goal `jscover-report:merge` (https://github.com/tntim96/JSCover-maven-plugin/issues/15)
 
 1.0.15 / 2015-08-26
 ==================